+++ pkg/ingester/ingester.go
@@ -1336,6 +1336,8 @@ func (i *Ingester) Push(ctx context.Context, req *cortexpb.WriteRequest) (*corte
 		failedHistogramsCount                  = 0
 		succeededExemplarsCount                = 0
 		failedExemplarsCount                   = 0
+		startTimestampSampleAppendFailCount    = 0
+		startTimestampHistogramAppendFailCount = 0
 		startAppend                            = time.Now()
 		sampleOutOfBoundsCount                 = 0
 		sampleOutOfOrderCount                  = 0
@@ -1460,6 +1462,14 @@ func (i *Ingester) Push(ctx context.Context, req *cortexpb.WriteRequest) (*corte
 		for _, s := range ts.Samples {
 			var err error
 
+			if s.StartTimestampMs != 0 && s.TimestampMs != 0 {
+				// TODO(SungJin1212): Change to AppendSTZeroSample after update the Prometheus v3.9.0+
+				if _, err = app.AppendCTZeroSample(ref, copiedLabels, s.TimestampMs, s.StartTimestampMs); err != nil && !errors.Is(err, storage.ErrOutOfOrderCT) {
+					startTimestampSampleAppendFailCount++
+					i.metrics.startTimestampFail.WithLabelValues(sampleMetricTypeFloat).Inc()
+				}
+			}
+
 			// If the cached reference exists, we try to use it.
 			if ref != 0 {
 				if _, err = app.Append(ref, copiedLabels, s.TimestampMs, s.Value); err == nil {
@@ -1504,6 +1514,14 @@ func (i *Ingester) Push(ctx context.Context, req *cortexpb.WriteRequest) (*corte
 					fh = cortexpb.FloatHistogramProtoToFloatHistogram(hp)
 				} else {
 					h = cortexpb.HistogramProtoToHistogram(hp)
+				}
+
+				if hp.StartTimestampMs != 0 && hp.TimestampMs != 0 {
+					// TODO(SungJin1212): Change to AppendHistogramSTZeroSample after update the Prometheus v3.9.0+
+					if _, err = app.AppendHistogramCTZeroSample(ref, copiedLabels, hp.TimestampMs, hp.StartTimestampMs, h, fh); err != nil && !errors.Is(err, storage.ErrOutOfOrderCT) {
+						startTimestampHistogramAppendFailCount++
+						i.metrics.startTimestampFail.WithLabelValues(sampleMetricTypeHistogram).Inc()
+					}
 				}
 
 				if ref != 0 {
@@ -1586,6 +1604,15 @@ func (i *Ingester) Push(ctx context.Context, req *cortexpb.WriteRequest) (*corte
 
 	// At this point all samples have been added to the appender, so we can track the time it took.
 	i.TSDBState.appenderAddDuration.Observe(time.Since(startAppend).Seconds())
+
+	if startTimestampSampleAppendFailCount > 0 || startTimestampHistogramAppendFailCount > 0 {
+		level.Debug(logutil.WithContext(ctx, i.logger)).Log(
+			"msg", "failed to append start timestamp in push",
+			"user", userID,
+			"sample_failures", startTimestampSampleAppendFailCount,
+			"histogram_failures", startTimestampHistogramAppendFailCount,
+		)
+	}
 
 	startCommit := time.Now()
 	if err := app.Commit(); err != nil {